+++ History.md
@@ -12,6 +12,7 @@
   * Added NX.Module.dropCollection追加
   * Added NX.Module.dropDatabase追加
   * Added NX.Module.listDatabase追加
+  * Added NX.Module.renameCollection追加